services/telegram-bot/requirements.txt
python-telegram-bot==22.0
pydantic==2.10.6
httpx==0.28.1
python-dotenv==1.0.1
loguru==0.7.3
datamodel-code-generator==0.28.4
